rename project to py-real-esrgan so we can add it to pypi
