+++ py_real_esrgan/rrdbnet_arch.py
@@ -2,7 +2,7 @@ import torch
 from torch import nn as nn
 from torch.nn import functional as F
 
-from real_esrgan.arch_utils import default_init_weights, make_layer, pixel_unshuffle
+from py_real_esrgan.arch_utils import default_init_weights, make_layer, pixel_unshuffle
 
 
 class ResidualDenseBlock(nn.Module):
