+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "Real-ESRGAN"
-version = "1.0.0"
+name = "py-real-esrgan"
+version = "2.0.0"
 requires-python = ">= 3.10"
 authors = [
     {name = "Sberbank AI, Xintao Wang"},
@@ -39,4 +39,4 @@ Issues = "https://github.com/ai-forever/Real-ESRGAN/issues"
 Changelog = "https://github.com/ai-forever/Real-ESRGAN/blob/main/CHANGELOG.md"
 
 [project.scripts]
-real_esrgan = "real_esrgan.__main__:main"
+py-real-esrgan = "py_real_esrgan.__main__:main"
